Cryptozoic VCC Notice2 August, 2019
The validity check of super node physical nodes has started at 7 a.m, 27 August, 2019, UTC time (10:00am Moscow time 15:00PM Beijing time). The results are as follows:
46 super nodes do not meet the rules of super node because of offline, low client version and so on. Therefore, their queue position of super node have been moved to the end of super node queue.
It is recommended that the 46 super nodes, which are moved to the end of the super node queue, should install the latest version of the client and properly configure parameters, and maintain the normal and continuous supply of node network and power. In the next node check or super node quota increase, They can become the valid super node again.